Fix non-roaming overlay not work issue.
Roaming type is not expected if carrier config has not started
to load when SST tries to update roaming state. And if service
state is not change after carrier config loaded, the roaming
type will keep to be improper.

To fix this issue, add a flag that indicates whether carrier
config loaded. And poll state done again after carrier config
loading finished.

Change-Id: Idcc0a5d7469fdcac98dd38311b655e5b478577d1
CRs-Fixed: 2344948